troca de lib serial
