+++ influunt-api/modules/influunt-device/app/os72c/client/device/SerialDeviceJava.java
@@ -1,33 +1,39 @@
 package os72c.client.device;
 
+import com.fazecast.jSerialComm.SerialPort;
+import com.fazecast.jSerialComm.SerialPortDataListener;
+import com.fazecast.jSerialComm.SerialPortEvent;
+import com.google.common.primitives.Bytes;
 import com.typesafe.config.Config;
 import engine.EventoMotor;
 import engine.IntervaloGrupoSemaforico;
 import engine.TipoEvento;
 import engine.TipoEventoControlador;
-import jssc.SerialPort;
-import jssc.SerialPortEvent;
-import jssc.SerialPortEventListener;
-import jssc.SerialPortException;
 import logger.InfluuntLogger;
 import logger.NivelLog;
 import logger.TipoLog;
 import models.TipoDetector;
-import org.apache.commons.lang.ArrayUtils;
 import org.apache.commons.math3.util.Pair;
 import org.joda.time.DateTime;
 import os72c.client.Client;
 import os72c.client.exceptions.HardwareFailureException;
-import play.Logger;
 import protocol.*;
 
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import static org.apache.commons.lang3.ArrayUtils.toArray;
 
 /**
  * Created by rodrigosol on 11/4/16.
  */
-public class SerialDevice implements DeviceBridge, SerialPortEventListener {
+public class SerialDeviceJava implements DeviceBridge, SerialPortDataListener {
 
 
     private final Config settings;
@@ -46,7 +52,7 @@ public class SerialDevice implements DeviceBridge, SerialPortEventListener {
 
     private DeviceBridgeCallback callback;
 
-    private SerialPort serialPort;
+    private com.fazecast.jSerialComm.SerialPort serialPort;
 
     private ScheduledExecutorService executor;
 
@@ -56,7 +62,9 @@ public class SerialDevice implements DeviceBridge, SerialPortEventListener {
 
     private int sequencia = 0;
 
-    public SerialDevice() {
+
+
+    public SerialDeviceJava() {
         settings = Client.getConfig().getConfig("serial");
         porta = settings.getString("porta");
         baudrate = settings.getInt("baudrate");
@@ -65,41 +73,43 @@ public class SerialDevice implements DeviceBridge, SerialPortEventListener {
         parity = settings.getInt("parity");
         startDelay = settings.getInt("startdelay");
 
-        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("Iniciando a municacao serial"));
-        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("PORTA    :%s", porta));
-        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("BAUDRATE :%d", baudrate));
-        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("DATABITS :%d", databits));
-        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("STOPBITS :%d", stopbits));
-        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("PARITY   :%d", parity));
-        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("START DELAY :%s", startDelay));
+        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("Iniciando a municacao serial"));
+        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("PORTA    :%s", porta));
+        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("BAUDRATE :%d", baudrate));
+        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("DATABITS :%d", databits));
+        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("STOPBITS :%d", stopbits));
+        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("PARITY   :%d", parity));
+        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("START DELAY :%s", startDelay));
     }
 
     public void start(DeviceBridgeCallback deviceBridgeCallback) {
         this.callback = deviceBridgeCallback;
         this.executor = Executors.newScheduledThreadPool(1);
 
-        serialPort = new SerialPort(porta);
-
+        serialPort = com.fazecast.jSerialComm.SerialPort.getCommPort(porta);
+        serialPort.setBaudRate(baudrate);
+        serialPort.setNumDataBits(databits);
+        serialPort.setNumStopBits(stopbits);
+        serialPort.setParity(parity);
         try {
-            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,"Abrindo a porta de comunicação");
+            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, "Abrindo a porta de comunicação");
             serialPort.openPort();//Open serial port
 
-            serialPort.setParams(baudrate, databits, stopbits, parity);
-            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,"Cumprindo delay");
+            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, "Cumprindo delay");
             Thread.sleep(startDelay);
-            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,"Limpando buffer");
-            serialPort.readBytes();
-            serialPort.purgePort(SerialPort.PURGE_RXCLEAR);
-            serialPort.purgePort(SerialPort.PURGE_TXCLEAR);
-            serialPort.readBytes();
-            serialPort.addEventListener(this);
-            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,"Comunicação serial pronta para iniciar");
-        } catch (SerialPortException spe) {
-            InfluuntLogger.log(NivelLog.NORMAL,TipoLog.ERRO,"Não foi possível iniciar a comunicação serial");
-            spe.printStackTrace();
-            throw new HardwareFailureException(spe.getMessage());
+            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, "Limpando buffer");
+            int bytesAvailable = serialPort.bytesAvailable();
+            if (bytesAvailable > 0) {
+                byte[] lixo = new byte[bytesAvailable];
+                serialPort.readBytes(lixo, bytesAvailable);
+            }
+            serialPort.addDataListener(this);
+            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, "Comunicação serial pronta para iniciar");
         } catch (Exception e) {
             e.printStackTrace();
+            InfluuntLogger.log(NivelLog.NORMAL, TipoLog.ERRO, "Não foi possível iniciar a comunicação serial");
+            throw new HardwareFailureException(e.getMessage());
+
         }
     }
 
@@ -116,10 +126,13 @@ public class SerialDevice implements DeviceBridge, SerialPortEventListener {
     private void send(Mensagem mensagem) {
         try {
             byte[] bytes = mensagem.toByteArray();
-            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("Enviando %d bytes pela serial",bytes.length));
-            serialPort.writeBytes(bytes);
-        } catch (SerialPortException e) {
-            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.ERRO,e.getMessage());
+
+            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("Enviando %d bytes pela serial", bytes.length));
+            //byte[] mgs = Bytes.concat("<I>".getBytes(), bytes, "<F>".getBytes());
+            //serialPort.writeBytes(mgs, mgs.length);
+            serialPort.writeBytes(bytes, bytes.length);
+        } catch (Exception e) {
+            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.ERRO, e.getMessage());
             e.printStackTrace();
         }
 
@@ -131,7 +144,7 @@ public class SerialDevice implements DeviceBridge, SerialPortEventListener {
         MensagemEstagio mensagem = new MensagemEstagio(TipoDeMensagemBaixoNivel.ESTAGIO, getSequencia(),
             intervaloGrupoSemaforico.quantidadeGruposSemaforicos());
         mensagem.addIntervalos(intervaloGrupoSemaforico);
-        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,mensagem.print());
+        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, mensagem.print());
         send(mensagem);
     }
 
@@ -141,26 +154,6 @@ public class SerialDevice implements DeviceBridge, SerialPortEventListener {
             sequencia = 0;
         }
         return ++sequencia;
-    }
-
-    @Override
-    public void serialEvent(SerialPortEvent serialPortEvent) {
-        if (serialPortEvent.isRXCHAR() && serialPortEvent.getEventValue() > 0) {//If data is available
-
-            int bytesCount = serialPortEvent.getEventValue();
-            try {
-
-                byte[] size = serialPort.readBytes(1);
-                byte[] complement = serialPort.readBytes(size[0] - 1);
-                byte[] msg = ArrayUtils.addAll(size, complement);
-                Mensagem mensagem = Mensagem.toMensagem(msg);
-                mensagemRecebida(mensagem);
-                InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,mensagem.getTipoMensagem().toString());
-            } catch (Exception e) {
-                InfluuntLogger.log(NivelLog.NORMAL,TipoLog.ERRO,e.getMessage());
-                e.printStackTrace();
-            }
-        }
     }
 
     private void mensagemRecebida(Mensagem mensagem) {
@@ -228,4 +221,46 @@ public class SerialDevice implements DeviceBridge, SerialPortEventListener {
     public void sendMensagem(TipoDeMensagemBaixoNivel inicio) {
         send(new MensagemInicio(TipoDeMensagemBaixoNivel.INICIO, getSequencia()));
     }
+
+    @Override
+    public int getListeningEvents() {
+         return SerialPort.LISTENING_EVENT_DATA_AVAILABLE;
+    }
+
+    @Override
+    public void serialEvent(SerialPortEvent event) {
+        if (event.getEventType() != com.fazecast.jSerialComm.SerialPort.LISTENING_EVENT_DATA_AVAILABLE) {
+            return;
+        }
+
+        StringBuffer buffer = new StringBuffer();
+
+        while(serialPort.bytesAvailable()>0 ){
+            byte[] newData = new byte[1];
+            int numRead = serialPort.readBytes(newData, 1);
+            buffer.append(new String(newData, StandardCharsets.US_ASCII));
+        }
+
+
+        Pattern pattern = Pattern.compile("<I>.*<F>");
+        Matcher matcher = pattern.matcher(buffer);
+
+
+
+        while(matcher.find()) {
+            Mensagem mensagem = null;
+            try {
+                mensagem = Mensagem.toMensagem(buffer.substring(matcher.start()+3,matcher.end()-3).getBytes(StandardCharsets.US_ASCII));
+                mensagemRecebida(mensagem);
+                InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, mensagem.getTipoMensagem().toString());
+
+            } catch (Exception e) {
+                InfluuntLogger.log(NivelLog.NORMAL, TipoLog.ERRO, e.getMessage());
+                e.printStackTrace();
+            }
+
+        }
+
+    }
 }
+
