+++ influunt-api/modules/influunt-device/app/os72c/client/device/SerialDeviceNativo.java
@@ -1,21 +1,24 @@
 package os72c.client.device;
 
-import com.fazecast.jSerialComm.SerialPort;
-import com.fazecast.jSerialComm.SerialPortDataListener;
-import com.fazecast.jSerialComm.SerialPortEvent;
 import com.typesafe.config.Config;
 import engine.EventoMotor;
 import engine.IntervaloGrupoSemaforico;
 import engine.TipoEvento;
 import engine.TipoEventoControlador;
+import jssc.SerialPort;
+import jssc.SerialPortEvent;
+import jssc.SerialPortEventListener;
+import jssc.SerialPortException;
 import logger.InfluuntLogger;
 import logger.NivelLog;
 import logger.TipoLog;
 import models.TipoDetector;
+import org.apache.commons.lang.ArrayUtils;
 import org.apache.commons.math3.util.Pair;
 import org.joda.time.DateTime;
 import os72c.client.Client;
 import os72c.client.exceptions.HardwareFailureException;
+import play.Logger;
 import protocol.*;
 
 import java.util.concurrent.Executors;
@@ -24,7 +27,7 @@ import java.util.concurrent.ScheduledExecutorService;
 /**
  * Created by rodrigosol on 11/4/16.
  */
-public class SerialDevice2 implements DeviceBridge, SerialPortDataListener {
+public class SerialDeviceNativo implements DeviceBridge, SerialPortEventListener {
 
 
     private final Config settings;
@@ -43,7 +46,7 @@ public class SerialDevice2 implements DeviceBridge, SerialPortDataListener {
 
     private DeviceBridgeCallback callback;
 
-    private com.fazecast.jSerialComm.SerialPort serialPort;
+    private SerialPort serialPort;
 
     private ScheduledExecutorService executor;
 
@@ -53,7 +56,7 @@ public class SerialDevice2 implements DeviceBridge, SerialPortDataListener {
 
     private int sequencia = 0;
 
-    public SerialDevice2() {
+    public SerialDeviceNativo() {
         settings = Client.getConfig().getConfig("serial");
         porta = settings.getString("porta");
         baudrate = settings.getInt("baudrate");
@@ -62,43 +65,41 @@ public class SerialDevice2 implements DeviceBridge, SerialPortDataListener {
         parity = settings.getInt("parity");
         startDelay = settings.getInt("startdelay");
 
-        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("Iniciando a municacao serial"));
-        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("PORTA    :%s", porta));
-        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("BAUDRATE :%d", baudrate));
-        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("DATABITS :%d", databits));
-        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("STOPBITS :%d", stopbits));
-        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("PARITY   :%d", parity));
-        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("START DELAY :%s", startDelay));
+        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("Iniciando a municacao serial"));
+        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("PORTA    :%s", porta));
+        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("BAUDRATE :%d", baudrate));
+        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("DATABITS :%d", databits));
+        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("STOPBITS :%d", stopbits));
+        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("PARITY   :%d", parity));
+        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("START DELAY :%s", startDelay));
     }
 
     public void start(DeviceBridgeCallback deviceBridgeCallback) {
         this.callback = deviceBridgeCallback;
         this.executor = Executors.newScheduledThreadPool(1);
 
-        serialPort = com.fazecast.jSerialComm.SerialPort.getCommPort(porta);
-        serialPort.setBaudRate(baudrate);
-        serialPort.setNumDataBits(databits);
-        serialPort.setNumStopBits(stopbits);
-        serialPort.setParity(parity);
+        serialPort = new SerialPort(porta);
+
         try {
-            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, "Abrindo a porta de comunicação");
+            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,"Abrindo a porta de comunicação");
             serialPort.openPort();//Open serial port
 
-            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, "Cumprindo delay");
+            serialPort.setParams(baudrate, databits, stopbits, parity);
+            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,"Cumprindo delay");
             Thread.sleep(startDelay);
-            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, "Limpando buffer");
-            int bytesAvailable = serialPort.bytesAvailable();
-            if (bytesAvailable > 0) {
-                byte[] lixo = new byte[bytesAvailable];
-                serialPort.readBytes(lixo, bytesAvailable);
-            }
-            serialPort.addDataListener(this);
-            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, "Comunicação serial pronta para iniciar");
+            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,"Limpando buffer");
+            serialPort.readBytes();
+            serialPort.purgePort(SerialPort.PURGE_RXCLEAR);
+            serialPort.purgePort(SerialPort.PURGE_TXCLEAR);
+            serialPort.readBytes();
+            serialPort.addEventListener(this);
+            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,"Comunicação serial pronta para iniciar");
+        } catch (SerialPortException spe) {
+            InfluuntLogger.log(NivelLog.NORMAL,TipoLog.ERRO,"Não foi possível iniciar a comunicação serial");
+            spe.printStackTrace();
+            throw new HardwareFailureException(spe.getMessage());
         } catch (Exception e) {
             e.printStackTrace();
-            InfluuntLogger.log(NivelLog.NORMAL, TipoLog.ERRO, "Não foi possível iniciar a comunicação serial");
-            throw new HardwareFailureException(e.getMessage());
-
         }
     }
 
@@ -115,10 +116,10 @@ public class SerialDevice2 implements DeviceBridge, SerialPortDataListener {
     private void send(Mensagem mensagem) {
         try {
             byte[] bytes = mensagem.toByteArray();
-            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, String.format("Enviando %d bytes pela serial", bytes.length));
-            serialPort.writeBytes(bytes, bytes.length);
-        } catch (Exception e) {
-            InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.ERRO, e.getMessage());
+            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,String.format("Enviando %d bytes pela serial",bytes.length));
+            serialPort.writeBytes(bytes);
+        } catch (SerialPortException e) {
+            InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.ERRO,e.getMessage());
             e.printStackTrace();
         }
 
@@ -130,7 +131,7 @@ public class SerialDevice2 implements DeviceBridge, SerialPortDataListener {
         MensagemEstagio mensagem = new MensagemEstagio(TipoDeMensagemBaixoNivel.ESTAGIO, getSequencia(),
             intervaloGrupoSemaforico.quantidadeGruposSemaforicos());
         mensagem.addIntervalos(intervaloGrupoSemaforico);
-        InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, mensagem.print());
+        InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,mensagem.print());
         send(mensagem);
     }
 
@@ -140,6 +141,26 @@ public class SerialDevice2 implements DeviceBridge, SerialPortDataListener {
             sequencia = 0;
         }
         return ++sequencia;
+    }
+
+    @Override
+    public void serialEvent(SerialPortEvent serialPortEvent) {
+        if (serialPortEvent.isRXCHAR() && serialPortEvent.getEventValue() > 0) {//If data is available
+
+            int bytesCount = serialPortEvent.getEventValue();
+            try {
+
+                byte[] size = serialPort.readBytes(1);
+                byte[] complement = serialPort.readBytes(size[0] - 1);
+                byte[] msg = ArrayUtils.addAll(size, complement);
+                Mensagem mensagem = Mensagem.toMensagem(msg);
+                mensagemRecebida(mensagem);
+                InfluuntLogger.log(NivelLog.DETALHADO,TipoLog.EXECUCAO,mensagem.getTipoMensagem().toString());
+            } catch (Exception e) {
+                InfluuntLogger.log(NivelLog.NORMAL,TipoLog.ERRO,e.getMessage());
+                e.printStackTrace();
+            }
+        }
     }
 
     private void mensagemRecebida(Mensagem mensagem) {
@@ -207,36 +228,4 @@ public class SerialDevice2 implements DeviceBridge, SerialPortDataListener {
     public void sendMensagem(TipoDeMensagemBaixoNivel inicio) {
         send(new MensagemInicio(TipoDeMensagemBaixoNivel.INICIO, getSequencia()));
     }
-
-    @Override
-    public int getListeningEvents() {
-         return SerialPort.LISTENING_EVENT_DATA_AVAILABLE;
-    }
-
-    @Override
-    public void serialEvent(SerialPortEvent event) {
-        if (event.getEventType() != com.fazecast.jSerialComm.SerialPort.LISTENING_EVENT_DATA_AVAILABLE) {
-            return;
-        }
-
-        byte[] size = new byte[1];
-        int numRead = serialPort.readBytes(size, 1);
-        if(numRead > 0 && size[0] > 0 && serialPort.bytesAvailable() >= size[0] ) {
-
-            byte[] msg = new byte[size[0]];
-            serialPort.readBytes(msg, msg.length);
-
-            Mensagem mensagem = null;
-            try {
-                mensagem = Mensagem.toMensagem(msg);
-                mensagemRecebida(mensagem);
-                InfluuntLogger.log(NivelLog.DETALHADO, TipoLog.EXECUCAO, mensagem.getTipoMensagem().toString());
-
-            } catch (Exception e) {
-                InfluuntLogger.log(NivelLog.NORMAL, TipoLog.ERRO, e.getMessage());
-                e.printStackTrace();
-            }
-        }
-    }
 }
-
