+++ influunt-api/modules/influunt-device/conf/application.conf
@@ -20,13 +20,13 @@
         # os72c.client.device.FakeDevice
         # Não comunicação com o hardware
         #
-        # os72c.client.device.SerialDevice
+        # os72c.client.device.SerialDeviceNativo
         # A hardware é conectado via protocolo serial
         #
         # os72c.client.virtual.ControladorForm
         # O sistema operacional se comunica com um controlador virtual
         #######################################
-        type = "os72c.client.device.SerialDevice2"
+        type = "os72c.client.device.SerialDeviceJava"
 
     }
